+++ src/components/menu/menu.styles.ts
@@ -6,13 +6,13 @@ export default css`
 
   :host {
     display: block;
-  }
-
-  .menu {
+    position: relative;
     background: var(--sl-panel-background-color);
     border: solid var(--sl-panel-border-width) var(--sl-panel-border-color);
     border-radius: var(--sl-border-radius-medium);
     padding: var(--sl-spacing-x-small) 0;
+    overflow: auto;
+    overscroll-behavior: none;
   }
 
   ::slotted(sl-divider) {